pyproject.toml
[tool.poetry]
name = "Kumiko"
version = "0.11.0"
description = "A multipurpose Discord bot built with freedom and choice in mind"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-dotenv = "^1.0.0"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
orjson = "^3.9.4"
asyncpraw = "^7.7.1"
uvloop = {markers = "sys_platform != \"win32\"", version = "^0.17.0"}
gql = { extras = ["aiohttp"], version = "^3.4.1" }
better-ipc = "^2.0.3"
redis = {extras = ["hiredis"], version = "^4.6.0"}
ciso8601 = "^2.3.0"
faust-cchardet = "^2.1.19"
discord-py = {extras = ["voice"], version = "^2.3.2"}
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus", rev = "8686b5d1bbc1d3c862292eb436ab630d6e9c9b53"}
asyncpg = "^0.28.0"
asyncpg-trek = "^0.3.1"
lru-dict = "^1.2.0"
psutil = "^5.9.5"
winloop = {markers = "sys_platform == \"win32\"", version = "^0.0.6"}
cysystemd = "^1.5.4"
python-dateutil = "^2.8.2"
msgspec = "^0.18.0"
langcodes = {extras = ["data"], version = "^3.3.0"}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
nox = "^2023.4.22"
nox-poetry = "^1.0.3"
pytest-cov = "^4.1.0"
dpytest = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pyinstrument = "^4.5.1"
pyright = "^1.1.322"
watchfiles = "^0.19.0"
jishaku = "^2.5.1"
ruff = "^0.0.284"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.1.2"
myst-parser = "^2.0.0"
sphinx-autobuild = "^2021.3.14"
furo = "^2023.7.26"
sphinxext-opengraph = "^0.8.2"
sphinx-copybutton = "^0.5.2"


[tool.isort]
profile = 'black'

[tool.pyright]
include = ["Bot/**"]
exclude = [
    "**/__pycache__", 
    "**/.mypy_cache", 
    "**/.dmpypy.json", 
    "Kumiko-Docs",
    "Docker"
]
reportMissingImports = true
typeCheckingMode = "basic"
reportUnnecessaryTypeIgnoreComment = "warning"

[tool.bandit]
skips = ["B311", "B101"]

[tool.ruff]
ignore = ["E501", "N999", "N801"]
select = ["E", "F", "N"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
